[circt-verilog] Add MapArithToComb pass to pipeline
Add the MapArithToComb pass to the circt-verilog pipeline. The
canonicalizers in the CF dialect produce `arith.select` and
`arith.constant` operations, but we would like the circt-verilog
pipeline to produce pure HW/Comb/Seq output where possible.

In the future we might want to move the pass from `Transforms` to
`Conversions`.