[FIRRTL] Refactor Linting pass, NFC
Refactor FIRRTL's linting pass to use a pattern where an object is created
which tracks state for the operation in question.  While not currently
used this way, I want to add state and this second object gives me some
place to put it without having to clear state in the actual pass itself.

This makes no functional changes to the code and just reogranizes existing
methods into this new class and then uses it in the pass.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>