[LLHD] Add loop unrolling pass (#8620)
Add an LLHD pass that unrolls loops with statically-known bounds.

The pass computes the dominance info for `llhd.combinational` ops, finds
loop back-edges by looking for blocks that branch back to blocks that
dominate them, which yields initial *loop head* and *back-edge* pairs.
Based on these the pass populates a data structure with loop metadata,
including the blocks in the loop body, exit control flow edges, the exit
condition, induction variable, and loop bounds. After gathering this
information, the loop body is simply stamped out muliple times, the few
relevant control flow edges are adjusted, and a rough cleanup of unused
blocks and values is performed.

The current implementation of the pass is *very* conservative. The loops
are required to have very simple control flow, and the induction
variable must go from 0 to a positive upper bound in increments of 1.
We'll want to relax these constraints in the future and allow other
increments/decrements, comparison predicates, and more.

Thanks again to @maerhart for doing a lot of initial work for this on
his experimental branch.

Co-authored-by: Martin Erhart <martin.erhart@sifive.com>