[LLHD] Make Deseq pass emit seq.firreg instead of seq.compreg (#8662)
The LLHD Deseq pass detects clocks and optional asynchronous resets. It
currently, mistakenly, creates `seq.compreg` ops with those clocks and
resets. The `seq.compreg` op turns out to have a _synchronous_ reset
though, making the lowering invalid. Use the `seq.firreg` op instead, as
it allows the reset type to be specified explicitly. This requries an
additional `comb.mux` to be created for potential enable signals, which
`seq.firreg` does not support directly.

We should really nail down the semantics of `seq.compreg` in a follow-up
commit. The documentation is very vague, and Deseq is likely not the
last pass mistakenly assuming the explicit reset of `seq.compreg` is
assumed to be asynchronous.