[VerifToSMT] Fix lowering of initial integer values for BMC (#8689)
Fixes a problem during VerifToSMT lowering when creating the smt::BVConstantOps for initial integer values of registers in the BMC operation. The current implementation calls getSExtValue() on the APInt provided by the initial value attribute. This causes a crash in either of the following situations:
- The initial value type is lager than 64 bits
- The initial value type is smaller than 64 bits and the MSB of the value is set. By sign-extending to 64 bits, the resulting integer value exceeds the width of the created BitVector.

This is fixed by passing the APInt directly to the builder of smt::BVConstantOp and adding a check beforehand ensuring that the types of the initial attribute and the initialized register match.


Co-authored-by: Martin Erhart <maerhart@outlook.com>