[Comb] Fix excessive const shifts causing crashes and invalid IR
The folders and canonicalizers for `comb.shl`, `comb.shru`, and
`comb.shrs` perform unchecked `getZExtValue()` calls which can crash
if the shift amount does not fit into 64 bits. The shift amount is also
not clamped to the width of the result, which causes the shift to be
replaced with ops that have a different type.

Found with [VlogHammer], which turns out to be a very handy tool!

Fixes #8694.
Fixes #8695.

[VlogHammer]: https://github.com/YosysHQ/VlogHammer