[hw] Convert HW Passes to use ODS constructors
Change all HW passes to use ODS consturctors as opposed to hand-rolling
them.  The hand-rolled constructors aren't necessary and add a bunch of
boilerplate.

Note: a number of these passes do not have users and hence may break out
of tree users.  The out of tree users are likely @teqdruid and @mortbopet.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>