[hw] Add container op
Add a new `hw.container` op.  This is intended to be a generic container
of hardware.  Critically, this exists to enable the checking of ops that
are users of inner symbol tables.  This is shown with the one error test
case that shows validation of a `hw.hierpath` op with an invalid inner
symbol use.  (This will not be checked if a `builtin.module` is used.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>