[FIRRTL] Do not allow uninferred widths or rests in enums
This adds verifiers to the FIRRTL parser and to the FEnumType to prevent
having uninferred widths and resets inside enumerations.  It seems like
the trickiness to implement these features would be hard to take
advantage of in practice, making it not worth it.  This also fixes the
error messages produced by the type verifier that had extraneous quotes
around the field name, and rearranges some tests to be co-located in the
test file.