[FIRRTL] FRT: support creating 0-valued enums (#8772)
This adds the ability to the FullResetTransform to create enumeration
values which are 0.  When the enumeration has a valid variant that is
encoded as a 0, we create this value using the FEnumCreateOp.  If there
is no valid variant with the value 0, we create a constant 0 and bitcast
it to the enumeration type.