[utils] Add Arcilator and Verilator comparison utility
Add the `compare-arcs-vtor.py` utility script which takes a core dialect
design, compiles it with Arcilator and Verilator to obtain two
simulators, and then runs the simulators in lockstep on random inputs
to verify that they produce the same output.

This can be useful in combination with circt-reduce in order to isolate
behavior mismatches between Arcilator and Verilator.