[ESI][Runtime] Windows wheel: use flat directory structure
Windows has no notion of rpath. It just looks in the same directory as
the executable/dll then the system search path. With the directory
structure as is currently was, `esiquery` was unable to load
ESICppRuntime.dll. It's not a great solution, but put everything in the
same directory on Windows.