[ImportVerilog] Bump slang to version 8
Update the Slang frontend to version 8.

This newer version of Slang supports building without RTTI and without
exceptions. We can get rid of the corresponding compiler flags now that
hacked up a no-RTTI and no-exceptions build before.

Any code in CIRCT that uses Slang's headers must now be built with the
C++20 option set, due to the headers exposing features of the standard
library that have been introduced in C++20. This only affects
ImportVerilog and the Verilog language server, which have additional
compile options set through the `SlangCompilerOptions.cmake` file
anyway.

It's a bit annoying that a few of Slang's dependencies are leaking
through, but that has been a long-standing issue. Now that we're closer
to being up-to-date with Slang, we might be able to get rid of these
eventually.

Co-authored-by: Tobias Wölfel <tobias.woelfel@mailbox.org>
Co-authored-by: Hideto Ueno <uenoku.tokotoko@gmail.com>
Co-authored-by: Fabian Schuiki <fabian@schuiki.ch>