[Reducer] Allow reductions to match multiple times per op
Instead of only allowing reductions to match or not match an entire op,
allow reductions to declare a list of zero, one, or multiple matches.
Each match has a separate `uint64_t` integer to identify it. This can be
an operand index for an operand forwarding reduction, or an argument
number for an argument removing reduction, or anything else.

This commit makes the reducer gather all of the potentially multiple
matches per op into a list. It then sorts this list by benefit and, as
before, uses the `rangeBase` and `rangeLength` to perform a binary
search through that list of matches. To apply reductions, the reducer
collects all matches on the same op and calls a new `rewriteMatches`
function with that list of matches. The reduction patterns can then
decide how to perform the rewrite with the given selection of matches.

In a subsequent PR, I will update a few of the HW reduction patterns to
use this new capability.