[HW] Make constantifier reduction treat results separately
Make the reduction pattern that replaces operation results with
constants more fine-grained. Since the reducer now supports multiple
matches per op, this reduction can generate a separate match for each of
the op results. This allows the reducer to replace _some_ of an op's
results with a constant, but not necessarily all of them.