[README] Frontload unified build to get started more easily; NFC
Adjust the build instructions part of the README such that it initially
mentions a very short snippet of code to compile CIRCT in a unified
build reasonably quickly. This assumes that the user has everything that
is needed, and they roughly know how git and cmake works.

This is then followed by the dependencies and how to get them, useful
build options, how to deal with submodules, and how to build LLVM/MLIR
separately from CIRCT. The idea is to move information relevant for most
users to the front, and keep information that only affects a few users
further to the back.

Also switch most of this to one-sentence-per-line to make diffs more
readable and reviewable.