[LLHD] Allow projections across blocks in mem2reg
The Mem2Reg pass currently refuses to promote projections that have uses
in different blocks. This is an unnecessary limitation, since the fact
that a projection dominates a drive means that any operands of the
projection, e.g. an array index, also dominate the drive. Therefore the
drive can still be replaced with extract/inject operations using the
projection operands. This only starts to become a problem once we allow
projections to be passed as branch destination operands, which we do
not.

Remove this unnecessary check, since it prevents some real-world use
cases from being properly handled by mem2reg.