[LLHD] Lower processes that have ops ahead of convergence block (#8856)
Currently, the LowerProcesses pass does not allow for the control flow
from the entry block and the wait op to contain any non-branch ops up
to the point where they converge. This is very conservative, since ops
like constants can make their way into the entry block in a way where
the entry block and wait op do converge, but with distinct yet
equivalent ops in the respective blocks leading up to the convergence
point.

Fix this issue by allowing blocks up to the convergence point to contain
any ops, as long as the block arguments passed to the convergence block
are coming from equivalent (but possibly distinct) ops.

We should really have a flavor of `llhd.process` that does not require
a wait and a control flow loop to describe combinational processes. That
would make these passes a lot easier.