[HW] Canonicalize array inject to create; allow struct inject loops
Add more canonicalization for inject ops. Array inject chains are now
traversed and converted to a single array create if all fields are set.
This also works in the presence of loops, which can occur in Verilog
input. Explicitly allow loops in struct injects, which can be broken by
injects if all fields are assigned.