[Comb] Generalize AndOp narrowing if mask has leading/trailing zeros
Instead of only narrowing `comb.and` operations that have a
`comb.extract` operand, do so on any and that has a constant operand
with leading or trailing zeros. Since we consider `comb.concat` to be
free, this canonicalization essentially turns a wide `comb.and` into a
cheaper narrow `comb.and` plus an additional, free `comb.concat`.

This helps canonicalize Comb ops that implement an integer equivalent
of `hw.array_inject` into corresponding extracts and concats.

In particular, this helps canonicalize
```mlir
hw.module @Foo(in %a : i1, out z : i2) {
  %c1_i2 = hw.constant 1 : i2
  %c-2_i2 = hw.constant -2 : i2
  %false = hw.constant false
  %0 = comb.and %5, %c-2_i2 : i2
  %1 = comb.concat %false, %a : i1, i1
  %2 = comb.or %0, %1 : i2
  %3 = comb.and %2, %c1_i2 : i2
  %4 = comb.concat %a, %false : i1, i1
  %5 = comb.or %3, %4 : i2
  hw.output %5 : i2
}
```
into the more compact
```mlir
hw.module @Foo(in %a : i1, out z : i2) {
  %false = hw.constant false
  %0 = comb.concat %false, %a : i1, i1
  %1 = comb.concat %a, %false : i1, i1
  %2 = comb.or %0, %1 : i2
  hw.output %2 : i2
}
```