[Synth] Add CutRewriter framework and TechMapper pass (#8868)
This commit implements a cut-based rewriting framework for combinational logic optimization 
and adds the TechMapper pass that uses NPN-equivalence matching with area and delay
metrics to map AIG circuits to technology libraries. The implementation is based on: 
* "Combinational and Sequential Mapping with Priority Cuts", Alan Mishchenko, Sungmin Cho, Satrajit Chatterjee and Robert Brayton, ICCAD 2007
* "Fast Boolean matching based on NPN classification",  Zheng Huang Lingli Wang Yakov Nasikovskiy and Alan Mishchenko, FPT 2013


The Technology Mapping Pass shows how to use this framework in practice by 
mapping generic logic operations to specific hardware gates from a technology 
library. This pass also serves as a test for the framework since 
testing the cut rewriting algorithms directly would be very difficult without 
a real application to exercise them.
The TechMapper pass processes `HWModuleOp` instances with `hw.techlib.info`
attributes as library patterns for now. This must be replaced with more well-designed representation.

The framework is designed to be easily extended with new optimization patterns 
and strategies while maintaining good performance for cut enumeration and 
pattern matching.