[ArcToLLVM] Disable pattern rollback in dialect conversion (#8878)
This PR sets the `allowPatternRollback` option to `false` in the ArcToLLVM and related dialect conversion passes. The main effect of this should be a reduction of compilation time and memory consumption. To the best of my knowledge we are not (naughtily) relying on the pre-conversion operations to be kept in the IR during conversion.