[MooreToCore] Convert unreachable, stop, finish, finish_message ops (#8895)
Add conversion patterns for the following ops:

- `moore.unreachable`
- `moore.builtin.stop`
- `moore.builtin.finish`
- `moore.builtin.finish_message`

Also create a `ConversionPatternSet` utility that simplifies creating
small conversion patterns. This basically allows us to define patterns
simply through a conversion function instead of having to define an
entire pattern struct. The `RewritePatternSet` already supports this,
but only for functions with the `matchAndRewrite` signature of rewrite
patterns:

    LogicalResult (Op, PatternRewriter&)

This new utility allows us to do the same for functions with the usual
`matchAndRewrite` signature of conversion patterns, too:

    LogicalResult (Op, Op::Adaptor, ConversionPatternRewriter&)