[ci] Release circt-{opt,synth,test} in firrtl-bin (#8928)
Add additional, useful tools as part of the `firrtl-bin` release package.
This will avoid users having to get one of the "full" installations (and
also deal with the fact that these installations are not truly static).

This has a cost of bloating the `firrtl-bin` tarball.  However, that seems
acceptable given that users may not have easy alternatives for how to get
these tools.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>