[Reduce] Use stable sort, prune unused symbols, tweak FIRRTL reductions
Use a stable sort in the reducer to sort matches. The binary search over
the matchese expects them to be in the order that they were matched in
the IR when matches have the same benefit.

Add a generic reduction to prune unused symbols.

Make the FIRRTL annotation remover check if annotations are already
empty, and not match on those ops. This significantly improves reduction
speed.

Fix an issue in the FIRRTL module externalizer, where `FExtModule`s
require an additional list of known layers that needs to be created.
Also, don't externalize modules that participate in NLAs, since removing
those would invalidate the NLA.