[ImportVerilog] Fix edge case for single-bit pre- and post-increment (#8945)
This PR fixes an edge case in UnaryExpression conversion from the Slang
AST to Moore. The current implementation maps pre- and post-increment
operations to binary additions of the target value and `1`. In this
step, the value `1` is assigned the same type as the target value it is
added to. In the edge case where the target value's type is an `i1`
(signed, width 1), this triggers an internal assertion, since a signed
single-bit integer may only represent the values {-1, 0}.

To handle this case, this PR introduces logic to first check whether the
target value is a single-bit integer. If so, the target value is
inverted, which corresponds to adding one without overflow saturation,
as prescribed by the pre- and post-increment operation. If it has a
different data type, the previous conversion is applied.

A test case for this pattern is added to
`tests/Conversion/ImportVerilog/basic.sv`.