[Arc] Add ExecuteOp
Add a new `arc.execute` op which is similar to `scf.execute_region`. It
allows an SSACFG region to be embedded in a graph region like an
`hw.module`. In a later PR, a conversion from `llhd.combinational` to
this new `arc.execute` will be added. We can't directly use
`scf.execute_region` for this since that op eagerly inlines into the
parent region even if the parent is a graph region. This does not
preserve semantics in the case of HW modules.