[Verif] handle self-referencing operations (#8972)
When cloning self-referencing operations like

    %reg = seq.firreg %reg clock %clock reset sync %reset, %c0_i4 : i4

lower-contracts would previously produce something like:

    ^bb0(%arg0: !seq.clock, %arg1: i1):
      ...
      %5 = seq.firreg(%5, %arg0, %arg1, %4) : i4
      ...

    verif.formal():
      ...
      %3 = seq.firreg(%5, %0, %1, %2) : i4
      ...

Here we can see that the verif.formal test refers to a value from `bb0`,
namely `%5`, which results in an error.

The problem appears to be that when cloning the operation, we haven't
yet created a mapping for the operation's result (`%5 -> %3`), so `%5`
is left in place.

This change makes us go through all operands of the cloned operations
once more, _after_ adding the mappings for the operations' results,
which means that the `%5` in verif.formal is remapped to `%3`.

This is done after cloning all operations to also handle cycles longer
than one operation.

Fixes #8952.