[Moore] Add builtins for simulation time measurements (#8970)
This PR adds builtins for `$time`, `$stime` and `$realtime` in the
`moore` dialect. All these builtins correspond to System Verilog system 
tasks and only slightly differ in their functionality.

To represent the return value of `$time`, this PR adds `TwoValueI64`, 
which represents a 64-Bit integer used to express the simulation time 
value in.