[SCFToCalyx] Adapt to upstream change in `LoopLikeOpInterface`.
This PR adapts to a change in `LoopLikeOpInterface` from
llvm/llvm-project#158679, which changed the API for getting a statically
known iteration count. The PR adapts to the change by using the accessor
of the `ForOp` for exactly that information instead of relying on a free
function that takes the properties/operands of the op.

Signed-off-by: Ingo Müller <ingomueller@google.com>